Add brand styling to your project

Add a _brand.yml and a logo.png to your project from DisplayR's inst/brands folder. If you need a new brand to be added, contact the Data Science Team

Usage

dr_add_brand(brand, directory = getwd())

Arguments

brand

Name of the brand you wish to load, as it appears in dr_list_brands()

directory

The directory you want the brand information to be added to, as a default it uses the current working directory

Value

Invisible

Examples

brands <- dr_list_brands()
dr_add_brand("shareds")